[Dashboard Agent] rename manual changes subscription
Align the dashboard attachment manual-sync wiring with the extracted subscription helper so mount sync composes the target structure more directly.

Made-with: Cursor
