+++ x-pack/platform/plugins/shared/dashboard_agent/public/attachment_types/on_attachment_mount.test.ts
@@ -141,7 +141,6 @@ describe('onAttachmentMount - manual changes sync', () => {
       checkSavedDashboardExist,
       getAttachment,
       updateOrigin,
-      addAttachment,
     } as unknown as OnAttachmentMountParams);
   };
